stable

+++ app/src/main/java/com/example/myapplication/UserBookInfo.kt
@@ -3,6 +3,6 @@ package com.example.myapplication
 data class UserBookInfo(
 
     val curPage: Int,
-    val curScroll:Int
+    val curScroll: Float
 
 )
